Refactor the flashing background code.
It would be ideal to be able to broadcast a message to all client live
views, channel style. Having to rely on the server state to manipulate
the DOM is not useful in this case...

Anyway, changed to a 500ms flash because network latencies can cause
problems since we rely on the server to clear the flash.