Handle the flashing background on the client
This makes sure that flashes last for 500ms, regardless of traffic on
the server or connection latency. It also cleans up the GameState
module, which is always welcome.

- use :blink because :flash is a reserved assign in Phoenix LV.
- send blink event using a data attribute in a div
  -> when the client picks up an update in this div, it adds a DOM node
     overlay to display the flash, the schedules a timer to remove it.
     It also sends back an event to the live view server to tell it that
     the blink event was processed so that the original state is restored.

It's a bit hackish, but seems to work well. It would be nicer to have a
"fire and forget" event being sent to the client.