Fix turn bug got odd number of players
When the two teams did not have the same number of players, the first
player of the team with less people played twice every full turn. This
is because the state was like this:

Given two teams : [a, b] and [c, d, e], the stack was
[a, b, a] + [c, d, e] = [a, c, b, d, a, e]

So, if we use it it gives the following

[a, c, b, d, a, e] [a, c, b, d, ...
 1  2  3  4  5  6   7  8  9  10 ...

We see that player a plays turn 5 *and* 7, when it should have been b
playing turn 7.

The refactored way of managing whose turn it is, is to represent teams
like this [[a, b], [c, d, e]], and the turns go like this
1 a plays, state becomes [[c, d, e], [b, a]]
2 c plays, state becomes [[b , a], [d, e, c]]
3 b plays, state becomes [[d, e, c], [a, b]]
...

It also opens up the game to have more than 2 teams.