Handle players spamming the "card guessed" button
When a player has a high latency, they can spam the "card guessed"
button, which results in multiple points scored for the team, each time
removing a card from the deck.

Now when a user click on the button, the card they currently see is sent
in the request payload. If this card doesn't match the current card to
be guessed, then we simply ignore the call.