Fix the "card guessed" spam fix
If a player has a high latency, and they send a :card_guessed at the
last second, then the game will tick, ending their turn, and changing
the current card to be guessed. So the guessed card sent by the player
will be different from the current card, but it will still be a legit
point.

Instead of comparing the card sent and the current card of the game, we
send the deck length seen by the player when they clicked "card
guessed". If this length matches the actual length of the deck when
receiving the message, then it's legit. Otherwise, the card has already
been guessed (thus removing it from the deck and changing its length),
and the point has already been given (we're actually using the index of
the card in the full deck as an id).