Fix the background blink on high latency clients
When a card was guessed or passed, the state was actually updated 3
times:

- once in the `handle_event("card_guessed", ...)` function, by using
  {:reply, assign(socket, game: game)}

- once in the `handle_event("update", ...)` because we broadcast it in
  `handle_event("card_guessed", ...)`

- once in the `handle_event("blink", ...)` because we broadcast it in
  `handle_event("card_guessed", ...)`

That's fine for low latency clients, but when the connection is slow,
those 3 different updates will modify the DOM  and sometimes causing weird
behaviour.

This modification of the state should actually be atomic and happen only
once, with one DOM update on the client. So all modifications of the
LiveView state actually happen in the `handle_event("update",...)`
callback. This might not be the most efficient, but it works quite well.

This was done in all the LiveView event handlers (those which modify the
state for all players and broadcast an "update" event).