Start GameServer GenServers when application starts
When a new version of the app is deployed, all existing GenServers are
terminated. Until now, it was the LiveView which restarted servers on
mount.

Now the GenServers are restarted by a Task when the application starts.
It makes it a bit easier to follow (if a game exists, then a GenServer
should exist, otherwise it's either an error, or the game is over and
has been cleaned up). It also avoids duplicating the "start if not
exists" code in the controller for the "join_game" action.

Only games that were updated in the last 24h are restarted on startup.
It could be much more aggressive (like the last 30min), but there's no
need for that right now given the amount of games being played.